build: switch from Makefile to Zig, add Windows support
- build.zig + build.zig.zon replace the Makefile. First build auto-fetches
  raylib 5.5 into deps/ (bypassing upstream's Zig 0.13/0.14-era build.zig)
  and pulls the sqlite amalgamation via the Zig package manager.
- Windows port: src/clipboard_win.c reads CF_DIB / PNG from the Win32
  clipboard; db.c/ui.c use %APPDATA% and %USERPROFILE%\Desktop; main.c
  loads Segoe UI when built for win32.
- paper_hash: do the mix in unsigned so Zig's debug UBSan stops tripping
  on signed-int overflow.

+++ src/canvas.c
@@ -11,8 +11,8 @@
 #define PAPER_TILE 256
 
 static unsigned int paper_hash(int x, int y) {
-    unsigned int h = (unsigned int)(x * 374761393 + y * 668265263);
-    h = (h ^ (h >> 13)) * 1274126177;
+    unsigned int h = (unsigned int)x * 374761393u + (unsigned int)y * 668265263u;
+    h = (h ^ (h >> 13)) * 1274126177u;
     return h ^ (h >> 16);
 }
